[#14] Common 파일 분리
